fix test; checkstyle/etc
