+++ processing/src/main/java/org/apache/druid/java/util/common/Cacheable.java
@@ -24,7 +24,7 @@ public interface Cacheable
   /**
    * Get a byte array used as a cache key.
    *
-   * @return a cache key
+   * @return bytes to be used as cache key - or null if this object should not be cached.
    */
   byte[] getCacheKey();
 }
