+++ processing/src/main/java/org/apache/druid/query/planning/ExecutionVertex.java
@@ -37,7 +37,6 @@ import org.apache.druid.segment.join.JoinPrefixUtils;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Stack;
 
 /**
  * Represents the native engine's execution vertex.
@@ -94,115 +93,6 @@ public class ExecutionVertex
                 .allMatch(ds -> ds instanceof TableDataSource));
   }
 
-  /**
-   * Execution vertex node.
-   *
-   * Can be a {@link Query} or a {@link DataSource}.
-   */
-  static class EVNode
-  {
-    final DataSource dataSource;
-    final Query<?> query;
-    final Integer index;
-
-    EVNode(DataSource dataSource, Integer index)
-    {
-      this.dataSource = dataSource;
-      this.query = null;
-      this.index = index;
-    }
-
-    EVNode(Query<?> query, Integer index)
-    {
-      this.dataSource = null;
-      this.query = query;
-      this.index = index;
-    }
-
-    public boolean isQuery()
-    {
-      return query != null;
-    }
-
-    public Query<?> getQuery()
-    {
-      Preconditions.checkNotNull(query, "query is null!");
-      return query;
-    }
-  }
-
-  static abstract class ExecutionVertexShuttle
-  {
-    protected Stack<EVNode> parents = new Stack<>();
-
-    protected final Query<?> traverse(Query<?> query)
-    {
-      try {
-        parents.push(new EVNode(query, null));
-        if (!mayTraverseQuery(query)) {
-          return query;
-        }
-        if (query instanceof BaseQuery<?>) {
-          BaseQuery<?> baseQuery = (BaseQuery<?>) query;
-          DataSource oldDataSource = baseQuery.getDataSource();
-          DataSource newDataSource = traverse(oldDataSource, null);
-          if (oldDataSource != newDataSource) {
-            query = baseQuery.withDataSource(newDataSource);
-          }
-        } else {
-          throw DruidException.defensive("Can't traverse a query[%s]!", query);
-        }
-        return visitQuery(query);
-      } finally {
-        parents.pop();
-      }
-    }
-
-    protected final DataSource traverse(DataSource dataSource, Integer index)
-    {
-      try {
-        parents.push(new EVNode(dataSource, index));
-        boolean traverse = mayTraverseDataSource(parents.peek());
-        if (dataSource instanceof QueryDataSource) {
-          QueryDataSource queryDataSource = (QueryDataSource) dataSource;
-          if (traverse) {
-            Query<?> oldQuery = queryDataSource.getQuery();
-            Query<?> newQuery = traverse(oldQuery);
-            if (oldQuery != newQuery) {
-              dataSource = new QueryDataSource(newQuery);
-            }
-          }
-          return visit(dataSource, !traverse);
-        } else {
-          List<DataSource> children = dataSource.getChildren();
-          List<DataSource> newChildren = new ArrayList<>();
-          boolean changed = false;
-          if (traverse) {
-            for (int i = 0; i < children.size(); i++) {
-              DataSource oldDS = children.get(i);
-              DataSource newDS = traverse(oldDS, i);
-              newChildren.add(newDS);
-              changed |= (oldDS != newDS);
-            }
-          }
-          DataSource newDataSource = changed ? dataSource.withChildren(newChildren) : dataSource;
-          return visit(newDataSource, !traverse);
-
-        }
-      } finally {
-        parents.pop();
-      }
-    }
-
-    protected abstract boolean mayTraverseQuery(Query<?> query);
-
-    protected abstract boolean mayTraverseDataSource(EVNode evNode);
-
-    protected abstract DataSource visit(DataSource dataSource, boolean leaf);
-
-    protected abstract Query<?> visitQuery(Query<?> query);
-  }
-
   static class ExecutionVertexExplorer extends ExecutionVertexShuttle
   {
     boolean discoveringBase = true;
@@ -253,11 +143,11 @@ public class ExecutionVertex
         discoveringBase = false;
       }
 
-      if (!leaf &&  dataSource instanceof JoinDataSource) {
+      if (!leaf && dataSource instanceof JoinDataSource) {
         JoinDataSource joinDataSource = (JoinDataSource) dataSource;
         joinPrefixes.add(joinDataSource.getRightPrefix());
       }
-      if(leaf && !isLeftLeaning()) {
+      if (leaf && !isLeftLeaning()) {
         allRightsAreGlobal &= dataSource.isGlobal();
       }
       return dataSource;
@@ -368,8 +258,8 @@ public class ExecutionVertex
     return new ReplaceBaseDataSource(baseDataSource, newBaseDataSource).traverse(topQuery);
   }
 
-  static class ReplaceBaseDataSource extends ExecutionVertexShuttle {
-
+  static class ReplaceBaseDataSource extends ExecutionVertexShuttle
+  {
     private DataSource newBaseDataSource;
     private DataSource oldBaseDataSource;
